+++ src/AdrPlus/Commands/UndoStatus/UndoStatusCommandHandler.cs
@@ -260,7 +260,7 @@ namespace AdrPlus.Commands.UndoStatus
                 var (_, Top) = _prompt.PromptCursorPosition();
                 DisplayWizardSummary(folderPrompt.Content, Path.GetFileName(filenewver.info.FileName));
                 var resultCnf = _prompt.PromptConfirm(Resources.AdrPlus.NewAdrPromptConfirmCreation, cancellationToken);
-                _prompt.PromptMovePosition(0, Top);
+                _prompt.PromptClearRegionFromTop(Top);
 
                 if (resultCnf.IsAborted)
                 {