+++ tests/AdrPlus.Tests/Commands/Version/VersionCommandHandlerTests.cs
@@ -2022,7 +2022,7 @@ namespace AdrPlus.Tests.Commands.Version
                 .Returns((false, false));
             var (_, Top) = ((0, 10));
             _mockConsole.PromptCursorPosition().Returns((0, Top));
-            _mockConsole.PromptMovePosition(0, Top);
+            _mockConsole.PromptClearRegionFromTop(Top);
 
             var repoConfig = new AdrPlusRepoConfig("", "")
             {